+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/room/OutboxMessageEventHandler.kt
@@ -64,7 +64,7 @@ class OutboxMessageEventHandler(
             } else null
         }
         if (removeOutboxMessages.isNotEmpty())
-            tm.transaction {
+            tm.writeTransaction {
                 removeOutboxMessages.forEach { roomOutboxMessageStore.update(it.first, it.second) { null } }
             }
     }