+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/user/ReceiptEventHandler.kt
@@ -29,7 +29,7 @@ class ReceiptEventHandler(
 
     internal suspend fun setReadReceipts(receiptEvents: List<ClientEvent<ReceiptEventContent>>) {
         if (receiptEvents.isNotEmpty())
-            tm.transaction {
+            tm.writeTransaction {
                 receiptEvents.forEach { receiptEvent ->
                     receiptEvent.roomIdOrNull?.let { roomId ->
                         log.trace { "set read receipts of room $roomId" }