+++ trixnity-client/src/commonMain/kotlin/net/folivo/trixnity/client/user/UserPresenceEventHandler.kt
@@ -2,9 +2,9 @@ package net.folivo.trixnity.client.user
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.datetime.Clock
-import net.folivo.trixnity.client.store.TransactionManager
 import net.folivo.trixnity.client.store.UserPresence
 import net.folivo.trixnity.client.store.UserPresenceStore
+import net.folivo.trixnity.client.store.TransactionManager
 import net.folivo.trixnity.clientserverapi.client.MatrixClientServerApiClient
 import net.folivo.trixnity.core.EventHandler
 import net.folivo.trixnity.core.model.events.ClientEvent
@@ -16,7 +16,7 @@ import kotlin.time.Duration.Companion.milliseconds
 
 class UserPresenceEventHandler(
     private val userPresenceStore: UserPresenceStore,
-    private val transactionManager: TransactionManager,
+    private val tm: TransactionManager,
     private val clock: Clock,
     private val api: MatrixClientServerApiClient,
 ) : EventHandler {
@@ -41,7 +41,7 @@ class UserPresenceEventHandler(
                     }
                 }
             }
-            transactionManager.transaction {
+            tm.writeTransaction {
                 newUserPresences.forEach { (userId, userPresence) ->
                     userPresenceStore.setPresence(userId, userPresence)
                 }