+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - User verification with user we have no direct room with, always creates an encrypted, trusted private, direct room
+- Fix race condition where the cache does no rollback, when no explicit transaction has been opened
 
 ### Security